Updating version string and removing files
I have changed the version string that is compiled into the code to
better reflect the way that git works: we now list the branch, the
latest tag and the SHA hash of the repository.  Also removed
datestamp.f90 and deps.obj.inc which should not have been in the
repository.