Bug fix for blips with NA, non-analytic integrals
There was an error in the loop over the grid for the non-local
projectors when using non-analytic blip integrals and the
neutral atom potential that caused an over-run.  Fixed with
a new loop.

Another small bug - the non-self-consistent exit from
new_SC_potl was commented out, and has been restored.