Moved dump_matrix2 from io_module2 to store_matrix_module
I am preparing store_matrix_module to keep the multiple information of matrix data
for multiple sets of atomic positions, for example during the MD.
The planned change will include the scheme to update the matrices on-memory,
without the IO files.
This is the first step : to do dump_matrix2 using the derived type "matrix_store".