Added warning for valence electron mismatch
If the number of valence electrons specified in the atom blocks
in Conquest_input is different to the number in the ion file, a
warning is printed.

I also changed the warning that is printed when L range is less
than S range so that it is only shown when performing O(N)
calculations (the only time it is important).