Bug fix for atom movement with O(N)
If O(N) is done with spin and moving atoms, we cannot at present
re-use the L matrix from step to step.  But this was breaking
relaxation, because L was zeroed at each step, and update_H was
calculating the density from L.  I've added a flag to stop this
when L isn't being re-used.