Adding L-matrix reuse with spin
This is a consolidation of three commits (original, expansion to
X and bug fix) for L  matrix re-use during MD, XL-BOMD and other
atom movement algorithms.  The original attempt used a development
branch as its head - this is fixing that.

Added save/load/rebuild of L matrix for both spin channels and
removed the abort if selected. (Originally contained in 9725804)

Also added save/load/rebuild for X, Xvel and related matrices
for both spin channels.  (Originally contained in 0a03467)

Fixed deallocation order for matrices in fmmi_XL which was
inconsistent with immi_XL when spin polarisation was turned
on.  (Originally contained in 1654719)