Introducing a minimum number of SCF iterations
The user can set this with SC.MinIters in the input file
(defaults to 2).  Introduced to address the issue that over
time during a relaxation an SCF cycle can drift leading to
a jump in force when a post-run static calculations is done.

Setting SC.MinIters to 3-5 can aid with this (and can be a good
idea in general to ensure convergence).