Issues with highly compressed PAOs, valence states
Credit to Kane Shenton for this: we found that for La the 4f states
are lower in energy than 5d even though they are unoccupied in the
pseudopotential (SG15) input.  Added a test for occupancy as well
as energy when re-ordering.

Also introduced a user-parameter to allow the semi-core energy
threshold to be set manually.