Tweak for non-perturbative polarisation PAOs
I changed the behaviour of the energy solver for polarisation
states because of Ti: the 4p state (polarisation) is actually
just bound, so we need to allow the lower bound on the energy
search to be negative.  Tested for Ti, Pb, Bi and La (all with
semi-core states, though not 4f for La).