Allow K matrix reuse after cell size change
Various changes to the Info and dump/grab_matrix2 routines so
that updating cell size is possible alongside reading in matrices.

The displacement vectors of neighbours are now fractional,
and atom_coord_diff is calculated on the basis of rescaled
atomic positions (i.e. if fractional coordinates do not
change, there is no change in position)