Bug fix for stress error with NA potentials
A misplaced gsum in NA projector force routines was causing the
NA stress to be wrong on multiple processors.  This has been
fixed.

I have also fixed an issue in H_matrix_module: NA matrix only
needs building once for each set of atomic positions (as with KE
and NL matrices) so I have moved the call to build it inside the
KE and NL if loop.  This is considerably more efficient for large
L and N values (particularly in the solid state).