Complete refactoring of XC code
Introduced a single call for XC potential (and dXC potential though
this is not yet implemented for LibXC), removed functional definitions
to XC module and created a simple switch to allow selection of code
that links against LibXC or not.

The only complication is that we have to keep the Conquest XC
functionals in XC_CQ_module.f90 and XC_LibXC_module.f90 in step,
but these should not be changed, so this won't be an obvious problem.

The final stage of the LibXC implementation is to provide routines
for the calculation of non-SCF forces (using get_dxc_potential).

I also tidied up the linking with LibXC - we now look for the F90
modules in the LibXC include directory meaning there's no need for
the source files themselves.