Non-SCF forces for LDA with LibXC
LibXC interface for the differential of the potential for LDA now
implemented, giving non-SCF forces and stresses with all LDA
functionals (spin and not, PCC and not).

Small changes: more comments and test of setting threshold for
zero density for better compatibility.

Final stage for LibXC implementation is GGA non-SCF forces.