Introduced check for consistency for cutoff
All radial tables are now tested to ensure that the cutoff and
the step size/number of points are consistent.  If not, then a
warning is printed and the cutoff adjusted to be (n-1)*step

The first two radial grid points are also tested for consistency
with the step size.  At present, a warning is printed but nothing
else done.