Added appropriate spin factors in non-spin areas
Several of the XC functionals in the Conquest XC routines are
non-spin-polaised only.  In the previous version of XC these
were passed the total density (i.e. scaled by the appropriate
spin factor).  Following the refactoring of the code to
interface with LibXC and to tidy the XC routines, this factor
was lost, giving significant errors.

spin_factor is now use-associated from global for the entire
XC module, and applied to the non-spin-polarised routines
correctly.