Implemented redistribution of atoms
If a process has an empty bundle (i.e. no responsibility for atoms)
then the space-filling curve routines for distributing atoms are
called, and atoms are redistributed dynamically.

This required small updates to fmmi and immi to ensure that all the
arrays and structures are deallocated and reallocated as the number
of partitions per process changes.