Implemented more details of MDMin
In the ASE MDMin implementation, they align the velocity along
the direction of the force at each step (this is done partly
by FIRE as well - though the proportion along the force is given
by the alpha parameter).

I have implemented a FIRE-like alpha (set with AtomMove.QMDAlpha)
to allow testing of the effectiveness of this.