Restore alpha=1 for L-BFGS and try new CG line minimisation
We use the backtracking line minimiser for both L-BFGS and CG.
There are good mathematical reasons to set alpha=1.0 for L-BFGS
but not for CG.  A simple heuristic approach has been set up:
if we satisfy the Armijo condition with the present value of
alpha then multiply by (1+scale) while simultaneously multiplying
scale by 0.9 (so that we don't increase too much).  Reasonably
effective in tests but needs more care.