Implement test for atoms too close to each other
This comes from the CheckInitialAtomicDistances branch in the
old repository, but I've moved the test to be much earlier in
the initialisation (saving a lot of time).

I've also changed mult_module.f90 to be Unix formatted instead
of DOS formatted, so the diff looks much bigger than it should
be.