Changed the code to make the velocity for the fixed atoms (components) zero explicitly.
Velocity for the fixed atoms (components) is now forced to be zero.
Reduced the degree of freedom by 3, if the centre of mass (COM) is fixed.
For now, it is assumed that FixedCOM is false, if we have some fixed atoms.
In the near future, we need to introduce a way to specify the atoms used in the definition of centre of mass.