Updates to SQNM
Mainly changes to deal with the situation where the energy rises
and we need to reset the search direction.  But associated with
that I have introduced new variables to track the change in energy
at the last electronic optimisation step (e.g. SCF or PAO) and
issue a warning to the user to tighten the tolerance.