Changed Pd electronic configuration for pseudopotential
The PseudoDojo valence configuration (4d10 as well as 4s and 4p
as semi-core states) makes the default basis set very poor.  Many
DFT codes have Pd as 4d9 5s1 (or 4d8 5s2 even) and we have changed
the pseudopotential to follow this.  The resulting default basis
is good quality.  We have added readme files to clarify the
change.