Unify the units in write_extxyz()
Since the distance unit in xyz file is Angstrom, we expect the units of force and energy are eV/Angstrom and eV.
However, the present units of force and energy are not. It is because the convert value of for_conv is 1.0 in default, then it will not change force unit when we use default units in Conquest_input.
Now, some local parameters of for_conv_loc, en_conv_loc, dist_conv_loc are used to
make all the units in this xyz file are with Angstrom (position) , eV (energy) and eV/Angstrom (force).