Bug fix for user-specified atoms
The order in which various parameters were read for perturbative
polarisation and user-specified atom blocks was wrong; in
particular the number of shells was read after polarisation
was tested giving an error for perturbative polarisation.

Also added code to remove basis set size when an atom
BasisBlock is specified (this potentially gave an error).