Further updates to stress
These changes are actually for non-SCF GGA calculations, but fall
under the general umbrella of stress bug fixes.

Note that the XC_GGA_stress (which comes from changes in the XC
energy related to cell-induced changes in the density gradient)
are not quite correctly implemented.  For non-SCF there should
be a mix of input and output densities (and Exc and Vxc) used
for this term, but this is very complex to implement, so for
now we have an approximation: we mix the stress contribution
found using input and output densities.  The proportion can
be set using an input flag (but should not really be changed
from 0.5)